AI/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_microservice"
version = "0.1"
description = "AI Microservice"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "fastapi==0.115.12",
    "uvicorn==0.34.3",
    "python-dotenv==1.1.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "httpx==0.28.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "praisonaiagents[llm]>=0.0.66",
    "mcp>=1.9.4",
    "beanie>=1.23.0",
    "bcrypt>=3.2.0,<4.0.0",
    "email-validator>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "motor>=3.3.0",
    "slowapi>=0.1.9",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.mypy]
ignore_missing_imports = true
mypy_path = ["app/infrastructure/stubs"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--color=yes"
]
markers = [
    "integration: marks tests as integration tests (require running server)",
    "slow: marks tests as slow running",
    "auth: marks tests related to authentication",
    "rate_limit: marks tests related to rate limiting"
]
asyncio_mode = "auto"